@yzhangcs 2025-01-18
@BBuf Hi, it looks `BLOCK_SIZE` does not actually work in this kernel?