@BBuf 2025-01-19
> @BBuf Hi, it looks `BLOCK_SIZE` does not actually work in this kernel?

Yeah, I'll fix it, thanks.